requirements_alignn.txt
-e git+https://github.com/usnistgov/alignn.git@229fe46c4647b123932a4d9981dcf5c68c18f96a#egg=alignn
ase==3.22.1
dgl>=0.6.0
ignite==0.5.0.dev20221024
jarvis-tools==2022.9.16
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.1
scikit_learn==1.1.3
tabulate==0.9.0
torch==1.13.1
tqdm==4.64.1
